requirements.txt
PyQt6==6.7.2
openai==1.35.13
anthropic==0.31.0
GitPython==3.1.43